+++ js/form-validation.js
@@ -1,0 +1,7 @@
+const checkLengthString = (string, length) => String(string).length <= length;
+
+export {
+  checkLengthString
+};
+
+console.log("Загрузился модуль form-validation.js");
